修复大部分问题

+++ src/main/java/me/matl114/hacks/modules/survival/BlockFarm.java
@@ -54,7 +54,7 @@ public class BlockFarm extends BaseModule {
    public final FlagRef whiteListEnable;
 
    public KalamaHelperHelperK<ItemStack> supplyItems(BlockItem blockItem) {
-      return InventoryUtils.v(s -> s.getItem() == blockItem ? -s.getCount() : null, true, false);
+      return InventoryUtils.v(s -> s.getItem() == blockItem ? (double)(-s.getCount()) : null, true, false);
    }
 
    public void onPlayerMineAttackBlock(Event<HitResult> hitResultEvent) {
